Create UserContext in Event page.
refs #1
