+++ src/App.vue
@@ -14,30 +14,11 @@
 <script lang="ts">
 import { defineComponent, onMounted } from "vue";
 import { useRoute } from "vue-router";
-import { UserContext } from "./models/user_context";
-import {
-  LocalStorageEventRepository,
-  LocalStorageCommentRepository,
-} from "./repository/local_storage";
-import {
-  FirebaseCommentRepository,
-  FirebaseEventRepository,
-} from "./repository/firebase";
 
 export default defineComponent({
   name: "App",
   setup() {
-    const createOrGetUserContext = () => {
-      if (localStorage.user_context) {
-        return UserContext.fromJSON(localStorage.user_context);
-      } else {
-        const userContext = new UserContext();
-        localStorage.setItem("user_context", userContext.toJSON());
-        return userContext;
-      }
-    };
-    const userContext = createOrGetUserContext();
-    return { userContext };
+    return {};
   },
 });
 </script>
